chore: close out audit — zero clippy, dedup grpc stack, test-isolation convention
Finishes the 2026-08-13 audit's remaining open items (§2.5, §2.7, §2.9).

§2.9 — clippy to zero across --workspace --all-targets. The mechanical
majority via --fix; the judgement calls by hand: two `Result<_, ()>` returns
became named error enums (ArenaError, LargeStackError, PathError); the
detour-trampoline transmutes were written out by type so the ABI assumption is
documented at the cast; the two NT-ABI trampolines carry a reasoned
`#[allow(too_many_arguments)]` (arity mirrors NtCreateFile).

§2.7 — duplicate dependency versions ~13 → 4. The bulk traced to tonic 0.12
pinning tower 0.4 and the old hashbrown/indexmap/socket2. Upgraded the gRPC
stack to tonic 0.14 / prost 0.14 (codegen moved to tonic-prost-build) and
thiserror 2, and bumped our windows-sys 0.59 → 0.61 to match clap/tokio. The
four that remain (syn 2+3, windows-sys 0.52) are pinned by proc-macro crates
and retour→region — not resolvable from this tree. Also silenced the benign
LNK4216 on vfs-payload (its entry point is reached by PE RVA, never by name).

§2.5 — the test-isolation pattern is now a stated convention at VA_LOCK: a
test asserting on process-global state takes a lock like it, or lives in its
own test binary. A sweep confirmed the only within-binary group touching a
process-global is fully guarded.

Docs: audit statuses updated to reflect §2.4/2.6/2.8 (e604fd5) and this batch.
328 tests pass; build and clippy clean.

+++ rust/crates/vfs-inject/src/pe.rs
@@ -136,7 +136,7 @@ pub fn map_image_from_pe_bytes_local(pe: &[u8]) -> Result<(*mut c_void, usize), 
             let ex_dir = opt + 112 + 3 * 8;
             if ex_dir + 8 <= img.len() {
                 let ex_rva = rd_u32(&img, ex_dir) as usize;
-                let ex_size = rd_u32(&img, ex_dir + 4) as u32;
+                let ex_size = rd_u32(&img, ex_dir + 4);
                 if ex_rva != 0 && ex_size >= 12 {
                     type RtlAddFunctionTableFn =
                         unsafe extern "system" fn(*const c_void, u32, u64) -> u8;
